+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.6.3] - 2024-03-06
+
 ### Fixed
   - Plugin/Nibabel: crash in Convert3Dto4D when removing merged files